REORG, split app into functional groups.
app initializes, main runs.
