+++ create_db.py
@@ -1,4 +1,4 @@
-from ummbNet import *
+from main import *
 
 db.drop_all()
 db.create_all()
